pyproject.toml
[tool.poetry]
name = "auto-highlighter-py"
version = "0.0.3"
description = "automatically generate clips from VODs!"
authors = ["porplax <saynemarsh9@gmail.com>"]
readme = "README.md"
packages = [{include = "highlighter"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.1"
typer = "^0.12.5"
opencv-python = "^4.10.0.84"
pillow = "^10.4.0"
pycparser = "^2.22"

[tool.poetry.scripts]
auto-highlighter = 'highlighter:cli'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
